feat: login route. redirect without displaying UI. remove Reka
