+++ client/package.json
@@ -48,7 +48,6 @@
     "luxon": "3.5.0",
     "nuxt-snackbar": "1.1.2",
     "pinia": "2.2.2",
-    "reka-ui": "2.2.0",
     "sortablejs": "1.15.3",
     "sortablejs-vue3": "1.2.11",
     "vue": "3.5.8"
