+++ client/utilities/url.ts
@@ -1,0 +1,3 @@
+export const LOGIN_PATH = '/login/'
+
+export const onLoginRoute = (path: string) => path.startsWith(LOGIN_PATH)
