+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "nuxt": "^3.17.1"
+  }
+}
